fix: address review feedback
- Drop dead version_node None guard; index directly (flow_data ↔ node_set 1:1)
- Extract _revert_pipeline_to_version to lower revert_to_version complexity
- Reword Pipeline.revert_to_version docstring
- Use alertify.confirm for the revert button, matching project convention
- Move revert view test to its own module (trims test_chatbot_views length)
