+++ apps/chatbots/tests/test_chatbot_views.py
@@ -782,26 +782,6 @@ def test_web_chat_widget_rendering(client, team_with_users):
 
 
 @pytest.mark.django_db()
-def test_revert_chatbot_version_view(client, team_with_users):
-    team = team_with_users
-    user = team.members.first()
-    user.user_permissions.add(Permission.objects.get(codename="change_experiment"))
-    client.force_login(user)
-    experiment = ExperimentFactory.create(team=team, name="Original", owner=user)
-    version = experiment.create_new_version(make_default=True)
-    experiment.name = "Modified"
-    experiment.save()
-
-    url = reverse("chatbots:revert-version", args=[team.slug, experiment.id, version.version_number])
-    response = client.post(url)
-
-    assert response.status_code == 302
-    assert response.url.endswith("#versions")
-    experiment.refresh_from_db()
-    assert experiment.name == "Original"
-
-
-@pytest.mark.django_db()
 def test_version_operation_status_polling(client, team_with_users):
     """The status endpoint reports any in-flight version operation, not just publish."""
     team = team_with_users
