+++ templates/experiments/components/experiment_version_actions.html
@@ -23,11 +23,16 @@
             onclick="clearModalAndOpen();">
       View Details
     </button>
-    <form method="post" action="{% url 'chatbots:revert-version' team_slug=team.slug experiment_id=record.working_version_id version_number=record.version_number %}" class="inline"
-          onsubmit="return confirm('Revert the working chatbot to v{{ record.version_number }}? This overwrites the current unpublished changes.');">
+    <form method="post" action="{% url 'chatbots:revert-version' team_slug=team.slug experiment_id=record.working_version_id version_number=record.version_number %}" class="inline">
       {% csrf_token %}
       <div class="tooltip" data-tip="Revert working version to v{{ record.version_number }}">
-        <button type="submit" class="btn btn-sm btn-outline btn-primary">
+        <button type="button" class="btn btn-sm btn-outline btn-primary"
+                onclick="const form = this.closest('form'); alertify.confirm().setting({
+                  title: 'Confirm',
+                  message: 'Revert the working chatbot to v{{ record.version_number }}? This overwrites the current unpublished changes.',
+                  transition: 'fade',
+                  onok: function() { form.submit(); },
+                }).show();">
           <i class="fas fa-rotate-left"></i> Revert
         </button>
       </div>
